Fix bug in profiling, kvf can be float (#325)
* Fix bug in profiling (#324)

* Format numbers using scientific notation, remove coord kwarg

* Allow kvf to be a float. Add docstring arg info for kvf.

* Remove unused var "coords"

* Update changelog

+++ src/wam2layers/config.py
@@ -326,7 +326,7 @@ class Config(BaseModel):
         periodic_boundary: true
     """
 
-    kvf: int
+    kvf: float
     """net to gross vertical flux multiplication parameter
 
     For example:
